media-thumbnail-sidecar/requirements.txt
mitmproxy==12.1.0
requests_toolbelt==0.10.1
pillow==9.5.0
markupsafe==2.1.1
urllib3<2.0
